Faster config system using getattr
By overloading `__getattr__` instead of `__getattribute__` in the
config system, we avoid throwing and catching AttributeErrors all
the time. Since `__getattr__` is only called if the attribute is
not found normally, it is actually what makes sense in our situation
since we were overloading `__getattribute__` to do exactly this.

Addresses #720.