Make PES more like other learning rules
BCM and Oja calculate their delta, then apply it to the transform.
PES skipped this step and just incremented the transform (or decoders)
directly. By refactoring PES to calculate the delta and then apply it,
we get two big benefits:

1. The order in which learning rules are applied no longer matters;
   the delta calculations are all done before the transform changes,
   since delta calcs are updates and the transform changes are incs.
2. The code for defining a delta signal and applying it to the
   transform can be refactored into the `LearningRule` builder,
   rather than being duplicated in each `LearningRuleType` builder.
   This should make learning rules easier to implement in the future.

As another minor benefit, it's now possible to probe the 'delta' of
each learning rule. This makes it easy to see exactly what each
each learning rule is doing, which is especially useful when you're
using multiple learning rules at the same time.

Also made a few other minor enhancements: better docstrings for
`SimOja` and `SimBCM`, and a note about ordering of connections in
the `Network` builder.