Fix Python 3.4-specific NumPy warnings
With NumPy >= 1.8 and Python 3.4, a bunch of extra warnings are
raised when we set the warnings filter to be 'always'. Specifically,
we 1) set size_in and size_out to be np.int64s instead of ints, and
2) try to index into the dimensions of EnsembleArray output nodes with
floats instead of ints. This fixes both of those problems, which
(thanks to these warnings now showing up) don't become errors later on!