Remove signatures from notebooks
As we realized in #833, notebook signatures are machine-dependent.
We might as well remove them all and keep them out (with a test)
so that they're not showing up in our diffs all the time.

`devscripts/clearoutputs.py` has been changed to remove signatures,
so it should be easy to meet this requirement when changing notebooks.