Add EnsembleArray.add_neuron_input/output
Previously, these both happened whenever `neuron_nodes=True` was
passed in. However, there are occasions when you only want a
neuron input (as occurred in `InputGatedMemory`) or only want a
neuron output (for plotting a raster, for example).
Additionally, adding these functions makes the `__init__` function
a bit more readable.

For the time being, I've kept `neuron_nodes` around as a shortcut
to calling `add_neuron_input` and `add_neuron_output`. However,
I've added a DeprecationWarning when it's set to True; we should
remember to remove `neuron_nodes` in Nengo 2.2.

I also redid the unit tests appropriately, and moved the
EA.dimensions property up in the file, as I tend to expect all
properties to be organized before all methods.