Add custom ValidationError exception
This is the first of several custom exceptions, which all inherit
from the base NengoException. The main reason for this change is to
make it possible for applications that import Nengo to easily catch
Nengo-specific exceptions in a single try/except block and to simplify
exception-handling of Nengo-based applications in general.

The ValidationError is raised, generally, when the `validate` method
of a Param instance fails. However, we also raise this exception when
other issues occur in setting the attributes of an object in its
constructor.

By default, the ValidationError only keeps some of the traceback for
simplicity. However, if full tracebacks are desired, then an RC option
was added to enable full tracebacks. E.g.:

[exceptions]
simplify = False