Move label and seed to base NengoObject
This simplifies the subclass code a little bit, and makes their
interfaces more obvious and consistent.

Since `Node.seed` currently affects nothing, but would seem to
a user like it should affect things like the Process seed if
the Node output is a Process, we raise an error if the user
tries to set `Node.seed`.