Improve accuracy of LIF model
The way we currently compute overshoot uses a linear interpolation
scheme, whereas the previous equation can be used to solve for the exact
amount of overshoot (by substituting v(0) = 1 and rearranging for t).

But to make this work well, we must also deal with another issue: we
currently handle partially remaining refractory periods by
multiplicatively scaling the voltage, which is yet again a linear
approximation. Instead, we can apply the same trick and substitute the
amount of leftover time as t into the first equation (basically using a
different time-step per neuron).

This makes the neuron model's spike rate invariant to dt, under the
assumptions that dt <= tau_ref and the input is constant across each
time-step (aside: we can in fact relax the first assumption with O(1)
additional processing, although the approach becomes more
complicated). We can thus think of it as a perfect discretization for
digital hardware. By 'perfect', I mean the spike count differs from the
expected count by < 1 at all times.

Some testing reveals that this reduces the rmse(true_rates * T,
sum_spikes) versus the current LIF by as much as 90% after T = 2 seconds
of simulation! This is because the current model will consistently spike
slightly faster or slightly slower to accumulate O(t) error (see
Efficient and Accurate Time-Stepping Schemes for Integrate-and-Fire
Neuronal Networks, JCNS 2001).

This improvement was discovered while identifying a somewhat serious
problem with drift during integration (thanks to Wilten Nicola). Any
small errors from maintaining the refractory_period will add up over
time, causing systematic drift in one direction.

The improved neuron model reduces drift almost to zero, and this same
qualitative result occurs over a wide range of seeds. So the main idea
here is that this gives us the ability to run simulations at a higher
time-step without sacrificing performance (normally you need to drop dt
to see the same reduction in drift).

But, one drawback that I am aware of is this model can slow down the
network by a non-negligible amount. Some tests on my machine with an
integrator of 1000 neurons gave a slowdown of ~20%. I think this is due
to the expm1 and log1p being applied to arrays.