Fixes for Windows testing
- Use try/except for get_terminal_size

When testing in Python 3 with py.test in Windows, the
stdout buffer is "detached", which makes it impossible to query
the width and height of the terminal. This traps the error and
uses a default fallback of 80.

- Make sure to write UTF-8 files

By default, Windows writes files with the cp1252 encoding, which
fails when writing some examples to a Python script. I guess Linux
and Mac OS X use a different encoding by default which works for these,
but in any case we should be sure to always write UTF-8 files.
Since IPython returns us Unicode anyway, this actually removes a bit of
code that we had to make these files ASCII for no real reason.