Handle differing pickle protocols in cache
This augments the current legacy version with the pickle protocol
in order to ensure that the cache can always be loaded properly,
and to ensure that we don't have a mix of cache files dumped with
different pickle protocols. This can happen sometimes with developers
who are switching between Python 2 and Python 3, but is very unlikely
to affect users who are most likely not switching Python versions
very often. When they do switch, I think it's prudent to wipe the
cache clean.