Simplify learning rule build functions
By moving `model.params[rule] = None` to the generic build function.
Note that this is deliberately done *before* calling model.build
on the learning rule type so that if the learning rule type has some
build artifacts that it wants to store, it can still do this b
overriding the default `None` set in build_learning_rule.