Simplify neuron type build functions
Previously, three of the build functions were essentially the same
because they didn't require any additional state. This adds a generic
builder for NeuronType instances, so subclasses that do not create an
explicit builder will fall back to this generic builder, which works
for RectifiedLinear, Sigmoid, and LIFRate at the moment. The remaining
build functions are unaffected.

Additionally, this commit fixes a possible subtle bug with using a
mutable data structure as a keyword argument. Now, we use None and
make a new list if necessary.