Get rid of on_add stuff. Add stimuli property.
This slightly breaks backwards compatibility because the old style
syntax for action related stuff (instantiating Cortical etc. directly
and passing in actions) will not work anymore if that object is not
immediately added to the container. If someone is trying to do that an
ObsoleteError will be raised and probably no one ever used this
"feature".

This also adds a stimuli property to Modules which is an spa.Input
object so that this has not to be created manually.