Cache fails gracefully if from the future
If a user has been using `nengo>=2.1.0`, their cache version includes
the pickle protocol. If they then switch back to `nengo<2.1`, the
cache will error as it's expecting a single integer. This allows the
cache to be "forward compatible", in that it will not crash when
switching to earlier versions. Instead it will silently invalidate
the cache.