Modify tests run based on options
If the user passes a --simulator or --ref-simulator argument, then
the frontend tests will be skipped, and only the backend tests
corresponding to the passed option will be run.

This also skips the `compare` tests rather than removing them from
the `items` list, which will print a message when using the `-rs`
option, and will show compare tests as "skipped" in the summary.