Only cache things with explicit mark.
Addresses #1054.

Code can change and we can't detect it, so we allow only things that
are marked as supporting fingerprints to be cached and only mark
objects that should be stable as supporting this.

This also removes the solver_fn from the fingerprint because it is not
needed in there. It should always be the same function which is part of
the builder (the solver itself is still part of the fingerprint of
course). This in turn allows to simplify the SolverMock in the test
code a little bit. (It actually is slightly misnamed as it mocks the
solver_fn and not a Solver.)

Co-authored by: Jan Gosmann <jgosmann@uwaterloo.ca>
Co-authored by: Daniel Rasmussen <daniel.rasmussen@appliedbrainresearch.com>