Ensure flake8 errors fail TravisCI tests
TravisCI can only inspect the return code of the last command
run in the `script` list, so in our big if/else chain, we need
to ensure that the last command returns what we consider to be
success or failure.

In the static case, we were not looking at the return value of
flake8. Now, we use && to use both return values.