Add a pull request template for better reviewing
This commit adds a file, .github/pull_request_template.md,
which will prepopulate the text field in all new pull requests.
The intention is to make it easier for PR creators to include all
the relevant details to make reviewing easier.

The template is based on a few sources, most notably
https://github.com/TalAter/open-source-templates
with changes based on our development practices.

It seemed prudent to also add a CONTRIBUTING.rst file, which is
kind of like README.rst in that it's a shorter version of what's
in the documentation, hopefully not repeating too much.
I rearranged the dev documentation a bit to tighten it up as well.