Fix memory leak in WeakKeyIDDictionary
WeakKeyIDDictionary was not removing values when the key was freed.

We create a weak reference to the key with a callback here that will
free the value when no strong references to the key are left. To be
able to do this we need the id of the key which will be retrieved
with the _ref2id dictionary. Note that weak reference use the hash
function of the object they are referencing. Thus, we have to use the
id of the weak reference to index that dictionary. Finally, to support
the del operation we need to be able to map the object id to the weak
reference which requires another dictionary _id2ref (instantiating
another weak ref would give us an object with a different id). This
dictionary has another important function: It stores strong references
the weak reference to ensure that the callback gets called (it would
not if the weak reference gets garbage collected beforehand and we
need to store ids as keys in the _ref2id dictionary).

Fixes #1089.

Also added a test for WeakKeyIDDictionary value garbage collection.