Make writing of cache index atomic.
If the process is killed while writing the pickle, we can't load it
anymore and lose all index information. Thus, we write it to another
file first, so we still have the old index in case the process gets
killed. Once we're done writing we rename the file (which should be
atomic on most file systems and operating systems, but will at least
be a much faster operation reducing the likelihood of ending up with
an invalid index).

Addresses #1097.