Move cache context management to builder
Previously, the cache was entered in Simulator.__init__. This was
causing the current nengo_ocl's Simulator to fail as it calls the
builder without entering the cache. There's no reason why the cache
couldn't be entered in build_network instead, so I moved it there
so that nengo_ocl doesn't have to modify its Simulator.