Flesh out releasing documentation.
The release process has changed a little bit since these docs
were written, so I've updated them to better reflect what we
do now. The major changes are to add an "announcing" stage for
writing up the release announcement, and to add a step for
testing Nengo OCL when releasing Nengo.