Implement backend-independent learning rule.
The chosen names are probably not the best ... it's hard to name
things.

Every backend can implement GenericRuleBoundSignals in the way its
appropriate for that backend. The learning rule function get access
to all the pre and post signals with it. This alleviates the need
for external connections or probes, so that the same GenericRule
object can be used on multiple connections.