Add documentation for config system and presets.
This involved updating the docstrings for Config, ClassParams,
and InstanceParams, as well as modifying the documentation itself.
The config example reference was moved from the network section
to the new config section.

I actually think that this tells a pretty decent story now, as the
config system is well justified by the frontend API, which justifies
the use of pre-built networks, since you see how complicated they
can get when you deal with the cascading defaults system.