Set ensemble params in test_set_eval_points
Since we're only setting two evaluation points, it is possible for
the randomly generated ensemble properties to make tuning curves
for which neither of the eval points result in any activity, which
we raise as a build error. This sets the encoders and intercepts
so that there's always some activity.

Fixes #1160.