Add n_neurons property to Network.
This required some changes to EnsembleArray, since it already has
an n_neurons attribute. This doesn't seem to be used anywhere in
tests or examples, so I renamed it to `n_neurons_per_ensemble`.
It can also be accessed through the `n_neurons` attribute of
any ensemble within the ensemble array.

Addresses #435.