Move builder tests to better locations.
`test/test_builder.py` made sense as a location for builder tests
when the builder was just a file, but now that it's a folder a
separate test folder makes more sense. This commit moves all the
tests that were in `test_builder.py` to better locations, either
`builder/tests/test_signal.py` or `tests/test_simulator.py`.