Make all networks subclasses with no `net` arg
This commit converts the all network functions into classes to
standardize the way in which we expect modelers to create reusable
networks. In using subclasses, it is no longer possible to handle
the `net` argument (which we suspect was rarely used) so it was
removed.

Any network that did not include an `output` node had an output node
added to further standardize how custom reusable networks should be
organized.

Additionally, this allows the Product and CircularConvolution networks
to raise DeprecationWarnings when accessing inputs using their old
names.

Co-authored-by: Eric Hunsberger <eric.hunsberger@appliedbrainresearch.com>