Allow Param.validate to change the value.
This allows it to cast the value into the right type. For example it is
convenient to assign a list to an NdarrayParam, but this list should
get converted to an array before it is stored. Trying to do this
conversion is also usually the most robust way to validate the input
argument. Prior to these changes it was required to override __set__
to do such things and usually duplicate some of the validation code.

This also adds some missing super calls to do the validation
implemented in the Parameter base class.