Coerce parameters instead of validate
Previously, we had two methods that Parameter subclasses could
override: validate, which validated the value being set, and
__set__, which would actually set the value, possibly making
slightly modifications if appropriate.

However, this was not a super clear cut distinction because the
NdarrayParam modified the value in validate, and the main reason
to override __set__ is to modify the value. Sometimes the best
way to determine if a value is valid is to try to coerce it to
the desired form of that value. So, rather than having both
methods, this commit merges them into a single `coerce` method.
__set__ should not need to be overridden.

This also adds a `check_type` method for the common operation of
checking that a value is of a certain type, and raising a
ValidationError if that's not the case.

Co-authored by: Jan Gosmann <jgosmann@uwaterloo.ca>