squash! first try at caching
Move it back into one file, which fixes the main issue that the
conda environment wasn't being activated before since it was
being activated in another script, which is launched in a subshell.
sourcing the script would have fixed it, but I think .travis.yml
isn't long enough to warrant separate files at the moment.