Add three RatesToSpikesNeuronTypes
This commit adds three neuron types, all of which are subclasses
of a new base class, `RatesToSpikesNeuronTypes`. As the name
suggests, all of these types take in a rate neuron type and emit
spikes based on the desired rate. The three types are
`RegularSpiking`, `StochasticSpiking`, and `PoissonSpiking`. See
their documentation for details.

All of these types take in an `amplitude` parameter to scale the
spikes, which is useful when training with NengoDL.

In the backend, several changes were made to support these new
types. The `state` in `SimNeurons` has been split up into state
which are signals (`state_sigs`) and other state (`state_extra`).
This was done in order to handle the RNG in the same way as
other neuron type state, but this could be generalized in the
future to other stateful objects. Currently, however, passing objects
other than `RandomState` instances in `state` raises a `BuildError`.

Fixes #579.

Co-authored-by: Eric Hunsberger <eric.hunsberger@appliedbrainresearch.com>