Refactor OpMergeOptimizer
There are lots of little changes in here, but in general the commit
rearranges things that already existed, with very few changes
to the underlying logic.

Highlights
----------

Broke the _OpMerger class into several separate functions and
classes to meaningfully group the methods and to reduce the number
of arguments that had to be passed into each method.

Now, the _OpMerger is mostly a registry for the checker/merger
functions, and contains the core logic for determining if an op can
be merged (which was previously scattered in a few different places).

New classes that made sense in the reorganization are:

- _SigMerger, which groups together the functions checking for
  the mergeability of signals
- _OpInfo which caches some information about ops
- _OpsToMerge, which caches some information about a list of ops
  that will be merged in an optimization pass
- _OpMergePass, which encapsulates all the state needed for an
  optimization pass. _OpMergePass is a callable class that is
  called once per iteration of the main `optimize` function
  (see below).

The `AbstractMerger` class and all subclasses have been removed,
and are now replaced with decorated functions.

The `_OpMerger.optimize` method is now a bare function in the
`optimizer.py` module. This function is the only things that
the Simulator needs to know about, so it makes sense to keep
it as simple as possible (i.e., Simulator shouldn't need to know
that we create all those objects under the hood).

Verification
------------

In order to verify that this refactored version does the same
optimization as before, I printed out the number of ops before
and after optimization for all models in the test suite. Because
the location of data in main memory is non-deterministic, the
resulting number of ops can vary slightly run-to-run, but in general
the number of ops post-optimization is always very close to the
number of ops before refactoring (sometimes yielding slightly fewer
ops, sometimes slightly more ops).

Speed
-----

The point of the optimizer, of course, is to speed up the simulations.
In all of the models that I tested, the speed of each simulation run
before and after refactoring is very similar.

I have not done any in-depth profiling, but the speed of the
optimization step itself is, I believe, slightly faster
post-refactoring. The main reason for this is that the refactored
_OpMergePass no longer requires `{op: op}` to be added to the
`op_replacements` dictionary, and instead modifies `dg` based on
only the merged ops. This reduces the number of times that we
iterate of all of the ops, which can become costly for large models.