Remove SupportsRCDefaultsMixin
The refactored optimizer instead adds a new argument to
nengo.Simulator. This means that `max_passes` is not currently
configurable, but I believe that the idea behind
`SupportsRCDefaultsMixin` has more implications to the rest of the
frontend codebase, and is sufficiently different from the rest of the
work in this  PR, so I would like to move it to another PR and
discuss it separately.