Disable pointless-statement warnings.
In theory useful, but produces a lot of false positives in the tests
and it seems too much clutter to disable it for each test module
individually.

There is one spot in spa.Vocabulary where some bad coding style is
pointed out, but this will be rewritten in the new SPA system anyways.