Fix undefined-loop-variable warnings.
In some places assigning the default value does not make the code more
correct because the list length is either known in advance or the code
path after the loop will not be executed when no loop iterations where
executed. So maybe it is better to disable the warning completely, but
there were also some true positives.