Make sigmoid tests deterministic
In one case of the sigmoid tests, the intercept was not defined,
so it would fail on occasion. This commit ensures that all cases
have an intercept defined (which means that we don't need to seed
the network, which would have been another fix for this).

While I was there, I refactored the test into two separate tests
to cut down on repeated code.