Remove defective links from index page.
We don't generate a module index because we manually list frontend
API classes (and don't include the documentation of the containing
modules).

The link to the search page is not required as the actual search
function is provided in the sidebar for the theme that we use. (It
could work better, in particular provide better suggestions when
typing, but it works.)

Fixes #1316.