Disable default validation for some params
Some parameters rely on attributes of their parent object
to perform the validation, so the validation cannot be performed
until the attribute value is set on that object.

Additionally, Ensemble.n_neurons and Ensemble.dimensions
no longer have a default value set, meaning that they cannot
be configured to have a default.