Update progress bars asynchronously.
This obsoletes the ProgressUpdater classes and related functions
as this where intended for synchronous updating. They were also
used to deal with IPython<3 storing all output leading to huge
memory consumption. Since IPython 3 this should not be the case any
more, thus there is no downside to doing regular updates.

This involved major refactoring and improvements to the rest
of the progress bar implementation. Specifically:

- Task names are part of Progress (the model) rather than
  ProgressBar (the view).
- Progress bar views are closed explicitly.
- Progress bars now have stages, which allow the same progress bar
  to display the progress for different stages of an overall task,
  while still displaying the total time in the end.

Co-authored-by: Daniel Rasmussen <daniel.rasmussen@appliedbrainresearch.com>
Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>