Switch back to pytest.warns
Rather than using our own implementation.
The previous problems (see #1011) seem to be resolved (see #1378).

Doing this requires that we disable pytest-xdist for Python 2.7
on TravisCI, as random warnings can be emitted before the warning
filter is installed. It seems that pytest-xdist does not cause
significant speedups on TravisCI anyway, so we disabled it entirely.