Add sample_every argument to trange.
This creates matching tranges for probes with a sample_every
parameter that is not a multiple of the simulator timestep.

Note that this essentially replaces the old dt argument.
For backwards compatibility reasons, it still exists
as a keyword argument, but will warn and use the logic
of sample_every.

Fixes #1368.