Rework current and rates NeuronType methods
This makes explicit that one can get multiple sample points out
of a single call to current and rates.

This attempts to codify the current behavior and add in
some consistency to get fewer errors.

We cater to the most common use cases of these methods even if
it does break backward compatibility for the least commonly used
cases.

Specifically we:

- Consider any 1-d input as a vector of samples
- Always return a 2-d vector of shape (n_samples, n_neurons)

The result of these changes is that most use cases (like those in
the builder) are not changed. However, the case that we used to
advertise in our docs as being the main use case -- having
(n_neurons,) inputs -- now must be supplied as a (1, n_neurons)
ndarray. However, despite this being what is listed in the docs,
we actually did it in very few cases, and I suspect few other
people have used these methods.

Co-authored-by: Aaron Voelker <arvoelke@gmail.com>