Another rework option for current / rates
In this rework, we instead cater to the most common use cases
of these methods even if it does break backward compatibility
for the least commonly used use cases.

Specifically we:

- Consider any 1-d input as a vector of samples
- Return a 1-d vector if we only had 1 sample

The result of these changes is that most use cases (like those in
the builder) are not changed. However, the case that we used to
advertise in our docs as being the main use case -- having
(n_neurons,) inputs -- now must be supplied as a (1, n_neurons)
ndarray. However, despite this being what is listed in the docs,
we actually did it in very few cases, and I suspect few other
people have used these methods.