Fix broken or redirected links
And add linkchecking to CI. The checks will run on all builds,
but not when actually building docs, as we assume that the links
will have been checked to be not broken on every other commit.

Note that one link cannot be checked because its SSL configuration
is not supported by requests, the Python library that Sphinx uses
to check links. Another project decided to have the link in but
not have it be clickable as a compromise, so I've done the same
here. See https://github.com/SciTools/cartopy/issues/978.