Fix warning for deprecated pytest feature
See https://docs.pytest.org/en/latest/mark.html#updating-code
for why this change was needed.

We opted for the iter_markers approach as we might want to
parametrize twice to test all combinations of multiple
parametrizations rather than testing a set number of values
for some variables.

Also added a test. To implement the test, I made a small change
to the Plotter class, as the filename of the plot was not
originally available to the test function. This change makes it
available with no other changes to functionality (hopefully).