Activate pylint signature-differs warning.
Usually the signature should not differ so that subclasses
can be used in place of the parent classes. For the
UniformHypersphere distribution the default argument of None
for d is not valid, but whether a TypeError or a ValidationError
error is raised when d is not given seems hardly critical.