Automatic reprs work with all FrozenObjects
The FrozenObject._argreprs and _argrepr_params methods were also
combined, as the only reason for them to be separate was to cache
the defaults. However, since all things are readonly in the
rozenObject, we can cache the _argreprs instead to remove the
need to have two separate methods.

This commit also adds tests for Synapse and Distribution reprs.

Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>