Add versioning to documentation
Every time a tagged release is built, that documentation will be
permanently added to the docs so that people can go back and refer to
the documentation for previous versions.  The published docs
themselves are updated every time a new commit is pushed to master,
same as before.  I also have it set up so that docs are built and
pushed to `gh-pages-test` whenever a non-master branch is built.  We
could get rid of that if we want, as it does add to the regular build
time.  But I've found it handy in NengoDL for those times when you
want to make sure the documentation will build/display properly
within the overall doc versioning framework (since it isn't as easy
to build a full version of the docs locally anymore).  An example of
that `gh-pages-test` output, for this branch, can be seen at
https://github.com/nengo/nengo/commit/4534b8ed7fbcf1fa714df758
if you want to see what the versioned docs look like.