Fix progress bar in Jupyter Lab
The JavaScript renderer has a higher precedence than the vDOM
renderer. Thus the updates of VdomOrHtmlProgressBar were executing
the JS updates and not the vDOM updates in recent Jupyter Lab
versions. By sending the JavaScript embeded in HTML, the same
renderers will be used for the update as for the initial display
which ensures consistency no matter which renderer has precedence.
(But the HTML renderer also has a lower precedence than the
vDOM renderer.)

Fixes #1499.