Skip unsupported tests instead of xfail
When a simulator does not support a test, skip it instead of
xfailing to help differentiate between unsupported and other
tests, and to make it more clear that unsupported tests are
not run when running the rest of the test suite.

Also added the ability to run unsupported tests as (correctly)
xfailed using the `--unsupported` command line option.
This is somewhat tested, though there are issues with the test
when passing the `--unsupported` flag due to how Matplotlib
is loaded when using the `runpytest` method. However, this
will at least make sure that the flag has an effect.

Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>