Use simple decorators for Solver docs
We now have a few simple decorators when Solver classes need to
reduce duplication of parameter and attribute lists, instead of
using the `DocstringInheritor`, which has been removed.

In implementing these decorators, it was helpful to visualize the
class inheritance structure, which has been added to the docs in
case it is helpful to users. This required the addition of graphviz
to the TravisCI job generating the documentation. I also wrote
a quick docstring for SolverParam and added it to the docs as well,
since it shows up in the inheritance diagram and isn't easy to
remove.

In reworking the docstrings, it seemed that some base classes
only existed to make it easier to share docstrings. Since we are
using these decorators for that now, I removed those base classes
and instead had the more specific type inherit from the general one.
I think all these relationships make sense.