Change default intercept range to -1, 0.9
This avoids a problematic situation in which gains can become
arbitrarily large when the intercept value approaches 1, as
1 is the point that we attempt to reach when determining the
gain and bias value to assign to each neuron. If we are trying
to reach maximum firing at 1 and we start with no firing at 0.99,
we're going to end up with a huge gain, which can cause issues
for several types of models.