Use pytest-rng for rng and seed fixtures
This changed seeds, so several tests needed to be adjusted to pass.

Unfortunately, it is not possible to do this in a backwards
compatible way because having `pytest-rng` installed conflicts
with fixtures exposed in conftest.py with the same name.
Since the fixture acts the same as it did before, it should be
easy for downstream repositories to switch.

`pytest-rng` also uses a wider seed space, which was causing issues
with `utils.numpy.maxint`. This value was not technically correct
on 64-bit architectures either, so this was changed to the `maxseed`,
which is more correct, and matches all the places in which `maxint`
was used in the codebase.