Guard against NaNs in model
This makes two changes to guard against zero-length encoders.
The first is to add an explicit BuildError if any encoders have
zero length. This is useful on its own because it is not
immediately obvious why one would have a NaN encoder unless they
know how Nengo deal with encoders under the hood.
The second is to explicitly disallow NaNs in any of our signals.

This commit also contains a change to the optimizer, which is what
surfaced the issue in the first place. When merging two DotIncs,
we check that the Signals do not change post-merge. However, if
one of those signals contain a NaN, the check fails because NaNs
are not equal to themselves in NumPy. The check was changed to
an allclose check with zero tolerances because it has an `equal_nan`
parameter to check that NaNs are in the same position.

np.empty could return us an array with NaNs, so we now initialize
all of the signals in these tests. We also switch to the simpler
initialization syntax for Signals with unspecified initial values.

Fixes #1543.