Increase minimum numpy version to 1.13
We also add CI tests for 1.16 and 1.17, and have all NumPy
compatibility CI jobs run only for `release-candidate-` branches,
so for every release but not for every commit.
Since this reduces the number of jobs normally run to five,
we remove the two-stage process and allow all jobs to run
at the same time.