Fix simulator pickling
The ensemble param order does matter, as some parameters use
`n_neurons` and `dimensions` to check array shapes. These were
not uncovered in our tests likely because they were randomly set
in a workable order.

Simulators that are being unpickled should not be optimized a
second time because signals will change and conflict with the
signals that are being unpickled.