Check max rates in Sigmoid
Check that max rates are below `1 / tau_ref` in Sigmoid.

Also do all the max rates checks (in `LIFRate`, too) as not erroring
if all rates are below the limit (rather than erroring if any rates
are above), since if any max rates happen to be NaN, this will now
trigger the error (since `np.nan < x` is false for all `x`).

Also do not allow `tau_ref=0` in Sigmoid.