Get newest scipy.signals functions
And move them to `_vendor/scipy/signal`. This both makes it clearer
that these functions come from Scipy, and makes it easier to update
them because they are ignored for static checks/coverage tests.

The version of Scipy used, and modifications made, are noted at the
top of each file.