Allow NeuronType to initialize state
This puts more control in the frontend NeuronType object. For
all of the neuron types in Nengo core, we no longer need any
backend changes (builder functions) for each neuron type.

The first main change required for custom neuron types is to add a
`state` dict that declares all of the state variables this
neuron type needs, and maps to a distribution reflecting its
default initial value. The `state` dict must include the spikes/rates
state (where 'spikes' is used for spiking neurons and 'rates' for
non-spiking). These state variables will be initialized in a new
method on the NeuronType base class, `make_state`, which can be
overridden in subclasses if custom behavior is needed.

The second change required for custom neuron types is to accept
an `initial_state` argument to override the default initial value
distributions to either a new distribution or to a numpy array of
samples. The `NeuronType` base class should handle validation
appropriately, so all that is needed is to call `super().__init__`.

While most state variables are initialized to zero, an important
exception is any `voltage` variable, which is initialized to a
random value between 0 and 1. This is a more realistic and useful
starting state for neuron voltages.

Some additional NeuronType changes:

- We rename step_math to step. This is backwards compatible for now,
  but step_math will be removed in Nengo 4.0.0.
- We make `probeable` a property that returns the `state` tuple.
- We add a `spiking` property that uses the `state` to determine if
  the neuron type is spiking or non-spiking.

We also disable the progress bar when running tests. While this is
a slight optimization (skips the time printing to stdout), it was
mostly to make it easier to debug builder / optimizer code.
With the progress bar, the `pdb` console is basically unusable
because the progress bar keeps making new lines while trying
to do debug statements.

Co-authored-by: Trevor Bekolay <tbekolay@gmail.com>