Support nengo.Default in all frozen objects
Previously, Default was only supported in NengoObject and
LearningRuleType. In order to support it in more FrozenObject
subclasses, we effectively wanted to add the SupportDefaultsMixin
to FrozenObject; however, we needed a slight change in logic in
FrozenObject, so we opted for a small amount of duplication to
allow us to remove SupportDefaultsMixin. The previous behavior
is now in NengoObject.__setattr__ and the slightly changed logic
is in FrozenObject.__setattr__.