Add state property to SimNeurons/SimProcess
To automate the process of mapping from indexes to signals.

This mimics the pre-refactoring state attribute (i.e., dereferences
the state indices into actual Signals). This was a code snippet that
I was repeating a lot when working with those ops, so made sense to
centralize it, and it also makes it easier to write code that is
cross-compatibile pre/post refactoring.