Separate RatesToSpikesNeuronType into its own op
Rather than having the base_type SimNeurons op effectively nested
inside the RatesToSpikesNeuronType SimNeurons op. This simplifies
the backend implementation of these ops (since the RatesToSpikes
op implementation is completely independent of the base_type
op).

This also simplifies how state is handled for RatesToSpikes, since
there is a clean separation between the state of the base type
and the state of the wrapper.

As a side-benefit, this makes the intermediate output (`rate_out`)
probeable.