Run isort on all files
Fixes some cyclic imports that resulted from rearranging the imports,
and updates to the latest nengo-bones, which adds a static check
that all imports are ordered with isort.

Right now we don't discriminate between normal files and those
in `nengo/_vendor`, but that is something we might want to
do in the future.